% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{BetaProfile}
\alias{BetaProfile}
\title{\code{BetaProfile}: compute}
\usage{
BetaProfile(x, q = seq(0, 2, 0.1), datatype = "abundance", se = TRUE,
  nboot = 30, conf = 0.95)
}
\arguments{
\item{x}{a list consist of N data.frame/matrix describing species-by-assemblage/plot abundance. Note that
the species in each element must exactly match including specpes order. Use \code{data(abundata)} to see data example.}

\item{q}{a numeric value or vector specifying the diversity order of Hill number.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}).}

\item{se}{a logical variable to calculate the bootstrap standard error and conf confidence interval.}

\item{nboot}{an integer specifying the number of replications.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}
}
\value{
a list consists of N elements, where N is the number of region. Each element is a list containing 5 tables of gamma. alpha and beta diversities,
and 2 types of dissimialrities (Sorensen and Jaccard), respectively.
}
\description{
\code{BetaProfile}: compute
}
\examples{
data(abundata2)
out <- BetaProfile(x = abundata2, q = seq(0,2,0.1), datatype = "abundance")
}
