% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{ggbeta_profile}
\alias{ggbeta_profile}
\title{\code{ggbeta_profile}: plot the outcome of \code{BetaProfile} based on the \code{ggiNEXT} function.}
\usage{
ggbeta_profile(output)
}
\arguments{
\item{x}{the outcome of \code{BetaProfile}}

\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (type = 1); sample completeness curve (type = 2);
coverage-based rarefaction/extrapolation curve (type = 3).}
}
\value{
a list containing 2 ggplot2 object, \code{$betaDiv} for beta diversity and \code{$diss} for dissimilarity
}
\description{
\code{ggbeta_profile}: plot the outcome of \code{BetaProfile} based on the \code{ggiNEXT} function.
}
\examples{
data(abundata2)
out <- BetaProfile(x = abundata)
}
