% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggiNEXTbeta}
\alias{ggiNEXTbeta}
\title{\code{ggiNEXTbeta}: plot the outcome of \code{iNEXTbeta}.}
\usage{
ggiNEXTbeta(
  x,
  type = c("Beta diversity", "Dissimilarity"),
  level = c("Taxonomic", "Phylogenetic", "Functional(tau)", "Functional(AUC)"),
  scale = "free",
  main = NULL,
  transp = 0.3,
  line_size = 1,
  point_size = 1,
  stroke = 2
)
}
\arguments{
\item{x}{the outcome of \code{iNEXTbeta}}

\item{type}{the plot type: plot of beta diversity (\code{type = "Beta diversity"}) or plot of dissimilarity (\code{type = "Dissimilarity"})}

\item{level}{the level of diversity computed. only used to decide the label of y axis.}

\item{scale}{the scale used in \code{facet_grid} in ggplot. Default is \code{'free'}}

\item{main}{a string describing the title of the plot.}

\item{transp}{the transparency of the confidence interval. Default is 0.3.}

\item{line_size}{the size of the lines. Default is 1.}

\item{point_size}{the size of the points. Default is 1.}

\item{stroke}{the size of the border of the points. Default is 2.}
}
\value{
a plot of the input outcome of \code{iNEXTbeta}.
}
\description{
\code{ggiNEXTbeta}: plot the outcome of \code{iNEXTbeta}.
}
