% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{iNEXT_beta}
\alias{iNEXT_beta}
\title{\code{iNEXT_beta}: compute the rarefaction and extrapolation of beta diveristy}
\usage{
iNEXT_beta(x, q = 0, datatype = "abundance", size = NULL,
  endpoint = NULL, knots = 40, se = TRUE, conf = 0.95,
  nboot = 30)
}
\arguments{
\item{x}{a list consist of N data.frame/matrix describing species-by-assemblage/plot abundance. Note that
the species in each element must exactly match including specpes order. Use \code{data(abundata)} to see data example.}

\item{q}{a numeric value or vector specifying the diversity order of Hill number.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}).}

\item{size}{an integer vector of sample sizes (number of individuals or sampling units) for which diversity estimates will be computed.
If NULL, then diversity estimates will be computed for those sample sizes determined by the specified/default endpoint and knots.}

\item{endpoint}{an integer specifying the sample size that is the endpoint for rarefaction/extrapolation.
If NULL, then endpoint = double reference sample size.}

\item{knots}{an integer specifying the number of equally-spaced knots (say K, default is 40) between size 1 and the endpoint;
each knot represents a particular sample size for which diversity estimate will be calculated. If the endpoint is smaller than the reference sample size,
then \code{iNEXT_beta()} computes only the rarefaction esimates for approximately K evenly spaced knots. If the endpoint is larger than the reference sample size,
then \code{iNEXT_beta()} computes rarefaction estimates for approximately K/2 evenly spaced knots between sample size 1 and the reference sample size,
and computes extrapolation estimates for approximately K/2 evenly spaced knots between the reference sample size and the endpoint.}

\item{se}{a logical variable to calculate the bootstrap standard error and conf confidence interval.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{nboot}{an integer specifying the number of replications.}
}
\value{
a list of 3 iNEXT objects, representing gamma, alpha and beta diversity respectively.
the profiles of all six classes of evenness indices listed in Chao and Ricotta (2019) Ecology paper.
}
\description{
\code{iNEXT_beta}: compute the rarefaction and extrapolation of beta diveristy
}
\examples{
data(abundata)
out <- iNEXT_beta(x = abundata, q = c(0,1,2), datatype = "abundance")
}
