% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{iNEXTbeta}
\alias{iNEXTbeta}
\title{\code{iNEXTbeta}: compute the rarefaction and extrapolation of beta diveristy.}
\usage{
iNEXTbeta(
  x,
  coverage_expected,
  data_type = c("abundance", "incidence_raw"),
  level = c("taxonomic", "phylogenetic", "functional"),
  nboot = 20,
  conf = 0.95,
  phy_tree = NULL,
  reftime = NULL,
  distance_matrix = NULL,
  tau_type = c("single", "AUC"),
  tau = NULL,
  cut_number = NULL
)
}
\arguments{
\item{x}{a list consist of N data.frame/matrix describing species-by-assemblage abundance, if the data_type is "abundance".
a list consist of N lists and each element of the N list is a list consist of T data.frame/matrix describing raw incidence of every assemblage, if the data_type is "incidence_raw".
Note that the species in each element must exactly match including species order.}

\item{coverage_expected}{a numeric vector of sample coverage for which diversity estimates will be computed. Each number should between 0 and 1.}

\item{data_type}{data type of input data: individual-based abundance data (\code{data_type = "abundance"}) or sampling-unit-based incidence data (\code{data_type = "incidence_raw"})}

\item{level}{the level of diversity to be computed.}

\item{nboot}{an integer specifying the number of replications for se estimating. Setting 0 to ignore se estimating.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{phy_tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. used if \code{level="phylogenetic"}.}

\item{reftime}{a positive value specifying the reference times for diversity computation. If \code{NULL},
then \code{reftime} is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}. used if \code{level="phylogenetic"}.}

\item{distance_matrix}{a matrix describing the functional distance for all observed species in the pooled assemblage. used if \code{level="functional"}.}

\item{tau_type}{the functional diversity should be computed at given tau value \code{tau_type="single"} or for AUC \code{tau_type="AUC"}.}

\item{tau}{a positvie number specifying the level of threshold distinctiveness. used if \code{level="functional" and tau_type="single"}.}

\item{cut_number}{a integer specifying the number of cut when computing the AUC of functional diversity. used if \code{level="functional" and tau_type="AUC"}.}
}
\value{
a list consists of N elements, where N is the number of region. Each element is a list containing 7 tables of gamma, alpha and beta diversities,
and 4 types of dissimialrities (1-CqN, 1-UqN, 1-SqN, 1-VqN), respectively.
}
\description{
\code{iNEXTbeta}: compute the rarefaction and extrapolation of beta diveristy.
}
