% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{ggiNEXT_beta}
\alias{ggiNEXT_beta}
\title{\code{ggiNEXT_beta}: plot the outcome of \code{iNEXT_beta} based on the \code{ggiNEXT} function.}
\usage{
ggiNEXT_beta(output, type = 1, goal = "Diversity")
}
\arguments{
\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (type = 1); sample completeness curve (type = 2);
coverage-based rarefaction/extrapolation curve (type = 3).}

\item{x}{the outcome of \code{iNEXT_beta}}
}
\value{
a list containing 2 ggplot2 object, \code{$betaDiv} for beta diversity and \code{$diss} for dissimilarity
}
\description{
\code{ggiNEXT_beta}: plot the outcome of \code{iNEXT_beta} based on the \code{ggiNEXT} function.
}
\examples{
data(abundata)
out <- iNEXT_beta(x = abundata, q = c(0,1,2), datatype = "abundance")
ggout1 <- ggiNEXT_beta(output = out, type = 1)
ggout2 <- ggiNEXT_beta(output = out, type = 2)
ggout3 <- ggiNEXT_beta(output = out, type = 3)
}
